wiroute: move ip-link methods into a separate class/file
